+++ lib/ibm1149.h
@@ -91,7 +91,7 @@ static const unsigned short ibm1149_2uni[256] =
 
 
 static int
-ibm1149_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+ibm1149_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   *pwc = (ucs4_t) ibm1149_2uni[c];
@@ -135,7 +135,7 @@ static const unsigned char ibm1149_page00[] = {
 };
 
 static int
-ibm1149_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+ibm1149_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
    unsigned char c = 0;
    if (wc >= 0x0 && wc < 0x0100)
